+++ src/SqlSchema.fs
@@ -12,7 +12,7 @@ module internal Patterns =
       then Some (List.tail [ for g in m.Groups -> g.Value ])
       else None
 
-type Column = { Name:string; ClrType: Type; DbType: DbType; IsPrimarKey:bool; }
+type Column = { Name:string; ClrType: Type; DbType: DbType; IsPrimarKey:bool; IsNullable:bool }
 type Relationship = { Name:string; PrimaryTable:string; PrimaryKey:string; ForeignTable:string; ForeignKey:string }
 type Direction = In | Out
 type SprocParam = { Name:string; ClrType:Type; DbType:DbType; Direction:Direction; MaxLength:int option; Ordinal:int }
@@ -20,6 +20,8 @@ type Sproc = {FullName:string; Params:SprocParam list; ReturnColumns:Column list
 
 type Table = { Schema: string; Name:string; Type:string }
     with 
+        // Note here the [].[] format is ONLY used internally.  Do not use this in queries; Different vendors have 
+        // different ways to qualify whitespace.
         member x.FullName = sprintf "[%s].[%s]" x.Schema x.Name
         static member FromFullName(fullName:string) = 
             match fullName with